Update README.md
This dataset consists of tv shows and movies available on Netflix as of 2019. The dataset is collected from Flixable which is a third-party Netflix search engine.

In 2018, they released an interesting report which shows that the number of TV shows on Netflix has nearly tripled since 2010. The streaming service’s number of movies has decreased by more than 2,000 titles since 2010, while its number of TV shows has nearly tripled. It will be interesting to explore what all other insights can be obtained from the same dataset.

Initially, we start with understanding the data set, then we clean the data to make the analysis ready.

explore the data and understand the behavior of the same.

we then prepare the data for creating clusters by various parameters wherein we remove stop words, white spaces numbers, etc so that we can get important words and based on that we shall form clusters.

later I used the silhouette method and means elbow method to find the optimal number of clusters and built a recommender system by cosine similarity and recommended the top 5 movies.